// src/app/models/interface/occupant.ts
export interface IOccupant {
    // id: string;
    occ_age?: string;
    occ_created_on?: string;
    occ_gender?: string;
    occ_id?: string;
    occ_name?: string;
    occ_sno?: string;
    occ_state?: string;
    usr_id?: string;
  }
  export interface IDeleteOccupant {
    occ_id: string;
  };